Rename wifi to wifi_8266 to avoid name conflict with system files
