+++ src/wifi_8266.cpp
@@ -36,7 +36,7 @@
 #include "ratgdo.h"
 #include "config.h"
 #include "softAP.h"
-#include "wifi.h"
+#include "wifi_8266.h"
 
 // Logger tag
 static const char *TAG = "ratgdo-wifi";
